// src/app/shared/models/employee.dto.ts
export class Employee {
  employeeId?: number;
  employeeName?: string;
  employeeEmail?: string;
  employeeLoginId?: string;
  employeeBirthDate?: Date;
  employeeTelephone?: string;
  employeeNameKana?: string;
  employeeLoginPassword?: string;
  departmentName?: string;
  departmentId?: string;
  certificationName?: string;
  score?: number;
  endDate?: Date;
  startDate?: Date;
  certificationId?: string;
  employeeRePassword?: string;
}